Split month switcher from lunar day grid
